Added core module. Updated dependencies. Moved model for POI to core.
