+++ app/build.gradle
@@ -51,6 +51,8 @@ dependencies {
     compile 'jp.wasabeef:picasso-transformations:2.1.0'
     // If you want to use the GPU Filters
     compile 'jp.co.cyberagent.android.gpuimage:gpuimage-library:1.4.1'
+
+    compile project(':core')
 }
 
 apply from: "${project.rootDir}/config/quality/quality.gradle"
