--- app/src/main/java/org/charmeck/trailofhistory/model/PointOfInterest.java
@@ -1,14 +1,0 @@
-package org.charmeck.trailofhistory.model;
-
-/**
- * Created by Trey Robinson on 3/8/16.
- */
-public class PointOfInterest {
-
-  public String uid;
-  public String name;
-  public String description;
-  public double latitude;
-  public double longitude;
-  public String imageUrl;
-}
