+++ core/build.gradle
@@ -1,0 +1,38 @@
+apply plugin: 'com.android.library'
+apply plugin: 'com.neenbedankt.android-apt'
+
+android {
+  compileSdkVersion 24
+  buildToolsVersion "24.0.2"
+
+  defaultConfig {
+    minSdkVersion 19
+    targetSdkVersion 24
+    versionCode 1
+    versionName "1.0"
+  }
+  buildTypes {
+    release {
+      minifyEnabled false
+      proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+    }
+  }
+}
+
+dependencies {
+  compile fileTree(dir: 'libs', include: ['*.jar'])
+  testCompile 'junit:junit:4.12'
+  compile 'com.android.support:appcompat-v7:24.2.0'
+  compile 'com.android.support:design:24.2.0'
+  compile 'com.google.firebase:firebase-core:9.4.0'
+  compile 'com.google.firebase:firebase-database:9.4.0'
+  compile 'com.google.firebase:firebase-storage:9.4.0'
+  compile 'com.google.firebase:firebase-crash:9.4.0'
+  compile 'com.google.firebase:firebase-auth:9.4.0'
+  compile 'com.bignerdranch.android:simple-item-decoration:1.0.0'
+  compile 'com.jakewharton:butterknife:8.2.0'
+  compile 'com.google.android.gms:play-services-maps:9.4.0'
+  apt 'com.jakewharton:butterknife-compiler:8.2.0'
+}
+
+apply from: "${project.rootDir}/config/quality/quality.gradle"
