+++ core/src/test/java/org/charmeck/trailofhistory/core/ExampleUnitTest.java
@@ -1,0 +1,14 @@
+package com.charmeck.trailofhistory.core;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+/**
+ * To work on unit tests, switch the Test Artifact in the Build Variants view.
+ */
+public class ExampleUnitTest {
+  @Test public void addition_isCorrect() throws Exception {
+    assertEquals(4, 2 + 2);
+  }
+}
