+++ ranger/src/main/java/org/charmeck/trailofhistory/ranger/MapActivity.java
@@ -20,7 +20,7 @@ import com.google.android.gms.maps.SupportMapFragment;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
-import org.charmeck.trailofhistory.ranger.model.PointOfInterest;
+import org.charmeck.trailofhistory.core.model.PointOfInterest;
 
 /**
  * @author Antoine Campbell
