+++ ranger/src/main/java/org/charmeck/trailofhistory/ranger/NewPointOfInterestActivity.java
@@ -14,7 +14,7 @@ import butterknife.OnClick;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import org.charmeck.trailofhistory.ranger.manager.PointOfInterestManager;
-import org.charmeck.trailofhistory.ranger.model.PointOfInterest;
+import org.charmeck.trailofhistory.core.model.PointOfInterest;
 
 public class NewPointOfInterestActivity extends AppCompatActivity {
