+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':ranger'
+include ':app', ':ranger', ':core'
